+++ app/Controllers/Admin/Penerbit.php
@@ -4,14 +4,17 @@ namespace App\Controllers\Admin;
 
 use App\Controllers\BaseController;
 use App\Models\PenerbitModel;
+use App\Models\BukuModel;
 
 class Penerbit extends BaseController
 {
     protected $penerbitModel;
+    protected $bukuModel;
 
     function __construct()
     {
-        $this->penerbitModel = new PenerbitModel;    
+        $this->penerbitModel = new PenerbitModel;  
+        $this->bukuModel = new BukuModel;  
     }
 
     public function index()
@@ -53,17 +56,23 @@ class Penerbit extends BaseController
 
     function save()
     {
-        if ($this->penerbitModel->save($this->request->getVar()) == true) {
-            session()->setFlashdata('session',[
+        $req = $this->request->getVar();
+        if ($this->penerbitModel->where('kode_penerbit',$req['kode_penerbit'])->countAllResults() > 0) {
+            session()->setFlashdata('kotakok',[
+                'status' => 'warning',
+                'title' => 'Gagal',
+                'message' => 'Kode Penerbit Sudah Digunakan'
+            ]);
+            return redirect()->to(base_url('pustakawan/penerbit/tambah'))->withInput();
+        }
+        if ($this->penerbitModel->save($req) == true) {
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'success',
                 'message'   => 'Penerbit Berhasil Ditambahkan'
             ]);
             return redirect()->to(base_url('pustakawan/penerbit'));
-        }elseif(!$this->validate($this->penerbitModel->getvalidationRules())){
-            session()->setFlashdata('errors',$this->validator);
-            return redirect()->to(base_url('pustakawan/penerbit/tambah'))->withInput();
         }else{
-            session()->setFlashdata('session',[
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'error',
                 'message'   => 'Penerbit Gagal Ditambahkan'
             ]);
@@ -92,21 +101,20 @@ class Penerbit extends BaseController
 
     function update()
     {
+
         $penerbit = $this->request->getVar();
+
         if ($this->penerbitModel->where('kode_penerbit', $penerbit['kode_penerbit'])->set([
             'nama_penerbit' => $penerbit['nama_penerbit']
         ])->update() == true
         ){
-            session()->setFlashdata('session',[
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'success',
                 'message'   => 'Edit Penerbit Berhasil'
             ]);
             return redirect()->to(base_url('pustakawan/penerbit'));
-        }elseif(!$this->validate($this->penerbitModel->getvalidationRules())){
-            session()->setFlashdata('errors',$this->validator);
-            return redirect()->to(base_url('pustakawan/penerbit/ubah'.$penerbit['kode_penerbit']))->withInput();
         }else{
-            session()->setFlashdata('session',[
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'error',
                 'message'   => 'Edit Penerbit Dagal'
             ]);
@@ -116,14 +124,23 @@ class Penerbit extends BaseController
 
     public function delete($kode_penerbit)
     {
+        $this->bukuModel->disableForeignKeyChecks();
+        if ($this->bukuModel->where('kode_penerbit',$kode_penerbit)->countAllResults() > 0) {
+            session()->setFlashdata('kotakok',[
+                'status' => 'warning',
+                'title' => 'Gagal',
+                'message' => 'Penerbit Masih Digunakan'
+            ]);
+            return redirect()->to(base_url('pustakawan/penerbit'));
+        }
         if ($this->penerbitModel->where('kode_penerbit',$kode_penerbit)->delete() == true) {
-            session()->setFlashdata('session',[
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'success',
                 'message'   => 'Penerbit Berhasi Dihapus'
             ]);
             return redirect()->to(base_url('pustakawan/penerbit'));
         }else{
-            session()->setFlashdata('session',[
+            session()->setFlashdata('pojokatas',[
                 'status'    => 'error',
                 'message'   => 'Penerbit Gagal Dihapus'
             ]);
